Initial Commit
Basic Items and Classes modules with Index data retrieval methods created
Application uses a 'pinky out' start up script I wrote for Node apps that allows you to leave a 'startup.ts' file in any controller module root, and it will initialize the module.  This enables your server to retrieve a fresh OAuth key as soon as it starts up.  This ensures that any first request made to the server has a higher chance of success without having to manage the async delay.

Technology Used:
- TypeScript compiling to ES6 as per the tsconfig.json
- Express for easy routes and API endpoint management with the flexibility of middleware
- Prettier/ESLint for code style enforcement and uniform structure.

This marks Version "1.0.0" of what is currently the wow-classic-dkp API